server/requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.5
django-cors-headers==2.1.0
django-extensions==1.9.1
djangorestframework==3.6.4
idna==2.6
oauthlib==2.0.4
psycopg2==2.7.3.1
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.10.0
tweepy==3.5.0
urllib3==1.22
